rename libvmaf cuda header, install
