+++ libvmaf/include/libvmaf/meson.build
@@ -7,6 +7,14 @@ version_h_target = configure_file(input: 'version.h.in',
                                   output: 'version.h',
                                   configuration: version_h_data)
 
+is_cuda_enabled = get_option('enable_cuda') == true
+
+platform_specific_headers = []
+
+if is_cuda_enabled
+  platform_specific_headers += 'vmaf_cuda.h'
+endif
+
 # install headers
 install_headers(['libvmaf.h',
                 'compute_vmaf.h',
@@ -14,4 +22,5 @@ install_headers(['libvmaf.h',
                 'model.h',
                 'picture.h'],
                 version_h_target,
+                platform_specific_headers,
                 subdir : 'libvmaf')
