+++ libvmaf/include/libvmaf/vmaf_cuda.h
