+++ libvmaf/test/test_ring_buffer.c
@@ -25,7 +25,7 @@
 
 #include "test.h"
 #include "ring_buffer.h"
-#include "libvmaf/cuda.h"
+#include "libvmaf/vmaf_cuda.h"
 #include "thread_pool.h"
 #include "cuda/common.h"
 #include "cuda/picture_cuda.h"
